+++ src/main/java/org/eclipse/keyple/card/calypso/crypto/legacysam/CommandReadSamParameters.java
@@ -12,21 +12,19 @@
 package org.eclipse.keyple.card.calypso.crypto.legacysam;
 
 import static org.eclipse.keyple.card.calypso.crypto.legacysam.DtoAdapters.*;
+import static org.eclipse.keyple.card.calypso.crypto.legacysam.LegacySamConstants.SAM_PARAMETERS_LENGTH;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 import org.eclipse.keyple.core.util.ApduUtil;
-import org.eclipse.keyple.core.util.ByteArrayUtil;
 import org.eclipse.keypop.card.ApduResponseApi;
 
 /**
- * Builds the Read Event Counter APDU command.
+ * Builds the Read Parameters APDU command.
  *
- * @since 0.1.0
+ * @since 0.9.0
  */
-final class CommandReadCounter extends Command {
+final class CommandReadSamParameters extends Command {
 
-  private final int counterFileRecordNumber;
   private static final Map<Integer, StatusProperties> STATUS_TABLE;
 
   static {
@@ -35,45 +33,34 @@ final class CommandReadCounter extends Command {
         0x6900,
         new StatusProperties(
             "An event counter cannot be incremented", CounterOverflowException.class));
-    m.put(0x6A00, new StatusProperties("Incorrect P2", IllegalParameterException.class));
+    m.put(0x6A00, new StatusProperties("Incorrect P1 or P2", IllegalParameterException.class));
     m.put(0x6200, new StatusProperties("Correct execution with warning: data not signed"));
     STATUS_TABLE = m;
   }
 
   /**
-   * Instantiate a new CmdSamReadEventCounter
+   * Instantiates a new CmdSamReadCeilings.
    *
    * @param context The command context.
-   * @param counterFileRecordNumber The number of the counter file record to read (in range [0..2].
-   * @since 0.1.0
+   * @since 0.9.0
    */
-  CommandReadCounter(CommandContextDto context, int counterFileRecordNumber) {
+  CommandReadSamParameters(CommandContextDto context) {
 
-    super(CommandRef.READ_EVENT_COUNTER, 48, context);
-    this.counterFileRecordNumber = counterFileRecordNumber;
+    super(CommandRef.READ_SAM_PARAMETERS, 48, context);
+
     byte cla = context.getTargetSam().getClassByte();
-    byte p2 = (byte) (0xE1 + counterFileRecordNumber);
+    byte p1 = 0x00;
+    byte p2 = (byte) (0xA0);
 
     setApduRequest(
         new ApduRequestAdapter(
-            ApduUtil.build(
-                cla, getCommandRef().getInstructionByte(), (byte) 0x00, p2, null, (byte) 0x00)));
-  }
-
-  /**
-   * Retrieves the record number of the counter file that will be modified by this command.
-   *
-   * @return An int.
-   * @since 0.3.0
-   */
-  int getCounterFileRecordNumber() {
-    return counterFileRecordNumber;
+            ApduUtil.build(cla, getCommandRef().getInstructionByte(), p1, p2, null, (byte) 0x00)));
   }
 
   /**
    * {@inheritDoc}
    *
-   * @since 0.1.0
+   * @since 0.9.0
    */
   @Override
   Map<Integer, StatusProperties> getStatusTable() {
@@ -83,7 +70,7 @@ final class CommandReadCounter extends Command {
   /**
    * {@inheritDoc}
    *
-   * @since 0.3.0
+   * @since 0.9.0
    */
   @Override
   void finalizeRequest() {
@@ -93,7 +80,7 @@ final class CommandReadCounter extends Command {
   /**
    * {@inheritDoc}
    *
-   * @since 0.3.0
+   * @since 0.9.0
    */
   @Override
   boolean isControlSamRequiredToFinalizeRequest() {
@@ -103,18 +90,13 @@ final class CommandReadCounter extends Command {
   /**
    * {@inheritDoc}
    *
-   * @since 0.3.0
+   * @since 0.9.0
    */
   @Override
   void parseResponse(ApduResponseApi apduResponse) throws CommandException {
     setResponseAndCheckStatus(apduResponse);
-    byte[] dataOut = apduResponse.getDataOut();
-    for (int i = 0; i < 9; i++) {
-      getContext()
-          .getTargetSam()
-          .putCounterValue(
-              (counterFileRecordNumber * 9) + i,
-              ByteArrayUtil.extractInt(dataOut, 8 + (3 * i), 3, false));
-    }
+    byte[] keyParameter = new byte[SAM_PARAMETERS_LENGTH];
+    System.arraycopy(apduResponse.getApdu(), 8, keyParameter, 0, SAM_PARAMETERS_LENGTH);
+    getContext().getTargetSam().setSamParameters(new SamParametersAdapter(keyParameter));
   }
 }