+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Keypop Calypso Crypto Legacy SAM API `0.6.0` -> `0.7.0`
 
 ## [0.8.0] - 2024-09-06
 ### Added